// utils/Table/interface.ts
export type TableColorScheme =
  | string
  | "white"
  | "blackAlpha"
  | "gray"
  | "grayBlue"
  | "red"
  | "orange"
  | "yellow"
  | "green"
  | "blue"
  | "cyan"
  | "purple"
  | "techPink"
  | "techPurple"
// ...
export interface FiltersEditorProps {
  colorScheme?: TableColorScheme
  filterOperator: FilterOperator
  columnFilters: FilterOptionsState
  columnsOption: { value: string; label: string }[]
  onAdd: () => void
  onDelete: (index: number, columnFilters: FilterOptions) => void
  onChange: (index: number, columnFilters: FilterOptions) => void
  onChangeOperator: (filterOperator: FilterOperator) => void
}

export type CustomFilterFn =
  | "equalTo"
  | "notEqualTo"
  | "contains"
  | "doesNotContain"
  | "lessThan"
  | "notLessThan"
  | "moreThan"
  | "notMoreThan"
  | "empty"
  | "notEmpty"
  | "before"
  | "after"

export type FilterFn = FilterFnOption<any> & CustomFilterFn

export type FilterOptions = {
  id: string
  value: any
  filterFn?: FilterFn
}

export type FilterOptionsState = FilterOptions[]

export type FilterOperator = "and" | "or"